Create "run" subcommand for running AIR
We also move code into submodule.
