+++ tools/cli/air-trace/src/main.rs
@@ -14,144 +14,24 @@
  * limitations under the License.
  */
 
-use air_test_utils::avm_runner::AVMRunner;
-use avm_server::CallResults;
+mod run;
 
-use anyhow::Context as _;
-use clap::Parser;
-use std::path::{Path, PathBuf};
-use tracing_subscriber::EnvFilter;
+use clap::{Parser, Subcommand};
 
-pub const AQUAVM_TRACING_ENV: &str = "WASM_LOG";
-const DEFAULT_DATA: &str = r#"{"trace":[],"streams":{},"version":"0.2.2","lcid":0,"r_streams":{"$nodes":{}}}"#;
+#[derive(Parser)]
+struct Command {
+    #[clap(subcommand)]
+    command: Subcomm,
+}
 
-#[derive(Parser, Debug)]
-struct Args {
-    #[clap(long)]
-    init_peer_id: Option<String>,
-    #[clap(long)]
-    current_peer_id: Option<String>,
-    #[clap(long)]
-    timestamp: Option<u64>,
-    #[clap(long)]
-    ttl: Option<u32>,
-    #[clap(long = "call_results")]
-    call_results_path: Option<PathBuf>,
-
-    #[clap(long)]
-    max_heap_size: Option<u64>,
-    #[clap(long, default_value = "info")]
-    tracing_params: String,
-
-    #[clap(
-        long = "runtime",
-        env = "AIR_WASM_RUNTIME_PATH",
-        default_value = "target/wasm32-wasi/release/air_interpreter_server.wasm"
-    )]
-    air_wasm_runtime_path: PathBuf,
-    #[clap(long = "prev_data")]
-    prev_data_path: Option<PathBuf>,
-    #[clap(long = "data")]
-    data_path: PathBuf,
-    #[clap(long = "script")]
-    air_script_path: Option<PathBuf>,
+#[derive(Subcommand)]
+enum Subcomm {
+    Run(crate::run::Args),
 }
 
 fn main() -> anyhow::Result<()> {
-    let args = Args::parse();
-
-    update_tracing_env(&args.tracing_params);
-    init_tracing();
-
-    let mut runner = create_avm_runner(&args).context("Failed to instantiate AIR runner")?;
-
-    let air_script =
-        read_air_script(args.air_script_path.as_deref()).context("failed to read AIR script")?;
-    let prev_data = match &args.prev_data_path {
-        None => DEFAULT_DATA.to_owned(),
-        Some(prev_data_path) => load_data(prev_data_path).context("failed to read prev_data")?,
-    };
-    let current_data = load_data(&args.data_path).context("failed to read data")?;
-
-    let init_peer_id = args.init_peer_id.unwrap_or_else(|| "some_id".to_owned());
-    let timestamp = args.timestamp.unwrap_or_default();
-    let ttl = args.ttl.unwrap_or(u32::MAX);
-    let call_results = read_call_results(args.call_results_path.as_deref())?;
-
-    let results = runner
-        .call(
-            air_script,
-            prev_data,
-            current_data,
-            init_peer_id,
-            timestamp,
-            ttl,
-            call_results,
-        )
-        .context("Failed to execute the script")?;
-
-    println!("{:?}", results);
-
-    Ok(())
-}
-
-fn update_tracing_env(tracing_params: &str) {
-    std::env::set_var("WASM_LOG", tracing_params);
-}
-
-fn init_tracing() {
-    use tracing_subscriber::fmt::format::FmtSpan;
-
-    tracing_subscriber::fmt()
-        .with_env_filter(EnvFilter::from_env(AQUAVM_TRACING_ENV))
-        .json()
-        .with_span_events(FmtSpan::NEW | FmtSpan::CLOSE)
-        .init();
-}
-
-fn load_data(data_path: &Path) -> anyhow::Result<String> {
-    Ok(std::fs::read_to_string(data_path)?)
-}
-
-fn read_air_script(air_input: Option<&Path>) -> anyhow::Result<String> {
-    use std::io::Read;
-
-    let air_script = match air_input {
-        Some(in_path) => std::fs::read_to_string(in_path)?,
-        None => {
-            let mut buffer = String::new();
-            let mut stdin = std::io::stdin().lock();
-
-            stdin.read_to_string(&mut buffer)?;
-            buffer
-        }
-    };
-
-    Ok(air_script)
-}
-
-fn read_call_results(call_results_path: Option<&Path>) -> anyhow::Result<CallResults> {
-    match call_results_path {
-        None => Ok(CallResults::default()),
-        Some(call_results_path) => {
-            let call_results_json =
-                load_data(call_results_path).context("failed to read call_results")?;
-            Ok(serde_json::from_str(&call_results_json)
-                .context("failed to parse call_results data")?)
-        }
+    let command = Command::parse();
+    match command.command {
+        Subcomm::Run(args) => crate::run::run(args),
     }
-}
-
-fn create_avm_runner(args: &Args) -> anyhow::Result<AVMRunner> {
-    let current_peer_id = args
-        .current_peer_id
-        .clone()
-        .unwrap_or_else(|| "some_id".to_owned());
-
-    Ok(AVMRunner::new(
-        args.air_wasm_runtime_path.clone(),
-        current_peer_id,
-        args.max_heap_size,
-        0,
-    )?)
 }
